Update events.js
According to the comment in the issue, the current version of TypeScript does not contain the correct definitions for HTMLDialogElement. Therefore, the // @ts-ignore comments were added to suppress type errors related to the showModal() and close() methods.

The proposed fix in the comment is to update to version 4.8.3 of TypeScript, which should contain the correct definitions for HTMLDialogElement. Once you have updated to this version or a later version, you can remove the // @ts-ignore comments and TypeScript should be able to enforce the correct types for these methods based on the updated definitions.